Refactor/get facets (#1)
* no html

* refactor 2000

* update refactored database

* more pointers

* bucket tests and start with numbers

* more pointers

* buckets

* minor improvements

* more refactor

* update

* more update

* better for now

* update refactoring branch

* minor change

* add readme

* use mock facets if to wide

* simple when using default facets

* a lot of changes

* update storage and memory

* update readme

* update readme

* more updates

+++ pkg/facet/sort_test.go
@@ -13,11 +13,11 @@ func TestPresortedSorting(t *testing.T) {
 		4,
 		5,
 	}
-	ids := map[int64]struct{}{4: {}, 2: {}, 1: {}, 3: {}}
+	ids := map[uint]struct{}{4: {}, 2: {}, 1: {}, 3: {}}
 
 	sortedIds := sortIndex.SortMap(ids, 10)
 
-	expected := []int64{1, 3, 2, 4}
+	expected := []uint{1, 3, 2, 4}
 	if !reflect.DeepEqual(sortedIds, expected) {
 		t.Errorf("Expected %v but got %v", expected, sortedIds)
 	}
